--HG--
branch : feature

+++ src/libs/vdxf/dxflib/dl_writer_ascii.cpp
@@ -79,7 +79,7 @@ void DL_WriterA::dxfReal(int gc, double value) const
     // Cut away those zeros at the end:
     bool dot = false;
     int end = -1;
-    for (quint32 i=0, sz = strlen(str); i<sz; ++i)
+    for (quint32 i=0, sz = static_cast<quint32>(strlen(str)); i<sz; ++i)
     {
         if (str[i]=='.')
         {
