+++ src/libs/vtools/dialogs/tools/dialogcubicbezierpath.cpp
@@ -229,8 +229,8 @@ void DialogCubicBezierPath::currentPointChanged(int index)
 
             if (not data->IsUnique(newPath.name()))
             {
-                newDuplicate = DNumber(newPath.name());
-                newPath.SetDuplicate(newDuplicate);
+                newDuplicate = static_cast<qint32>(DNumber(newPath.name()));
+                newPath.SetDuplicate(static_cast<quint32>(newDuplicate));
             }
 
             ui->lineEditSplPathName->setText(newPath.name());
